+++ Wbbj/run_LH_PY_scalevar_onWbbj7.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# Py=yes: do PY shower and hadroniz.
+#PY=yes
+PY=yes
+
 dir=/home/oleari/fortran/POWHEG-BOX-V2/Wbbj
 
 
@@ -21,33 +25,33 @@ do
 	\rm pwgLHE*.top	 > /dev/null  2>&1
 	../runLHEF-scalevar.sh 
 
-	echo '  ***  doing PY6 shower  ***'
-	\rm pwg-*POWHEG+PYTHIA-output*.top  > /dev/null  2>&1
-	../runpy-scalevar.sh 
-
-	echo '  ***  doing PY6 shower + hadronization  ***'
-	\rm pwg-*POWHEG+PYTHIA-output*.top  > /dev/null  2>&1
-	../runpy-scalevar_had.sh 
-
+	if [ $PY = 'yes' ] 
+	then 
+	    echo '  ***  doing PY6 shower  ***'
+	    \rm pwg-*POWHEG+PYTHIA-output*.top  > /dev/null  2>&1
+	    ../runpy-scalevar.sh 
+	    
+	    echo '  ***  doing PY6 shower + hadronization  ***'
+	    \rm pwg-*POWHEG+PYTHIA-output*.top  > /dev/null  2>&1
+	    ../runpy-scalevar_had.sh 
+	fi
 	cd ../
     done
     cd ${work}pm-scalevar
     echo " "
     echo 'working in' $PWD  
-#    (echo 3; ls -c1 ../${work}m-scalevar/pwgNLO.top  ../${work}p-scalevar/pwgNLO.top; echo "") | mergedata
-#    mv fort.12 pwgNLO_Wpm.top    
-
     for i in {1..7}
     do
 	(echo 3; ls -c1 ../${work}m-scalevar/pwgLHEF_analysis-W${i}.top  ../${work}p-scalevar/pwgLHEF_analysis-W${i}.top; echo "") | mergedata
 	mv fort.12 pwgLHEF_analysis-W${i}_Wpm.top       
+	if [ $PY = 'yes' ] 
+	then
+	    (echo 3; ls -c1 ../${work}m-scalevar/pwgPOWHEG+PYTHIA-output_nohad-W${i}.top  ../${work}p-scalevar/pwgPOWHEG+PYTHIA-output_nohad-W${i}.top; echo "") | mergedata
+	    mv fort.12 pwgPOWHEG+PYTHIA-output_nohad-W${i}_Wpm.top       
 	
-	(echo 3; ls -c1 ../${work}m-scalevar/pwgPOWHEG+PYTHIA-output_nohad-W${i}.top  ../${work}p-scalevar/pwgPOWHEG+PYTHIA-output_nohad-W${i}.top; echo "") | mergedata
-	mv fort.12 pwgPOWHEG+PYTHIA-output_nohad-W${i}_Wpm.top       
-	
-	(echo 3; ls -c1 ../${work}m-scalevar/pwgPOWHEG+PYTHIA-output_had-W${i}.top  ../${work}p-scalevar/pwgPOWHEG+PYTHIA-output_had-W${i}.top; echo "") | mergedata
-	mv fort.12 pwgPOWHEG+PYTHIA-output_had-W${i}_Wpm.top       
-
+	    (echo 3; ls -c1 ../${work}m-scalevar/pwgPOWHEG+PYTHIA-output_had-W${i}.top  ../${work}p-scalevar/pwgPOWHEG+PYTHIA-output_had-W${i}.top; echo "") | mergedata
+	    mv fort.12 pwgPOWHEG+PYTHIA-output_had-W${i}_Wpm.top       
+	fi
 
     done
 
@@ -61,12 +65,14 @@ do
 	(echo $ind; ls -c1 pwgLHEF_analysis-W*_Wpm.top; echo "") | mergedata
 	mv fort.12 pwgLHEF_analysis-${which}_Wpm.top 
 
-
-	(echo $ind; ls -c1 pwgPOWHEG+PYTHIA-output_nohad-W*_Wpm.top; echo "") | mergedata
-	mv fort.12 pwgPOWHEG+PYTHIA-output_nohad-${which}_Wpm.top 
-
-	(echo $ind; ls -c1 pwgPOWHEG+PYTHIA-output_had-W*_Wpm.top; echo "") | mergedata
-	mv fort.12 pwgPOWHEG+PYTHIA-output_had-${which}_Wpm.top 
+	if [ $PY = 'yes' ] 
+	then
+	    (echo $ind; ls -c1 pwgPOWHEG+PYTHIA-output_nohad-W*_Wpm.top; echo "") | mergedata
+	    mv fort.12 pwgPOWHEG+PYTHIA-output_nohad-${which}_Wpm.top 
+	    
+	    (echo $ind; ls -c1 pwgPOWHEG+PYTHIA-output_had-W*_Wpm.top; echo "") | mergedata
+	    mv fort.12 pwgPOWHEG+PYTHIA-output_had-${which}_Wpm.top 
+	fi
 
     done
 	
@@ -76,12 +82,17 @@ dir=/home/oleari/fortran/POWHEG-BOX-V2/Wbbj/Wbb_Wbbj_results/Wbbj/Wbbj7_minlo_nf
 
 \cp pwgLHEF_analysis-min_Wpm.top                 $dir
 \cp pwgLHEF_analysis-max_Wpm.top                 $dir
+\cp pwgLHEF_analysis-W1_Wpm.top                  $dir/pwgLHEF_analysis_Wpm.top
 
-\cp pwgPOWHEG+PYTHIA-output_nohad-min_Wpm.top    $dir
-\cp pwgPOWHEG+PYTHIA-output_nohad-max_Wpm.top    $dir
+if [ $PY = 'yes' ] 
+then
+    \cp pwgPOWHEG+PYTHIA-output_nohad-min_Wpm.top    $dir
+    \cp pwgPOWHEG+PYTHIA-output_nohad-max_Wpm.top    $dir
+    \cp pwgPOWHEG+PYTHIA-output_nohad-W1_Wpm.top     $dir/pwgPOWHEG+PYTHIA-output_nohad_Wpm.top
 
-\cp pwgPOWHEG+PYTHIA-output_had-min_Wpm.top      $dir
-\cp pwgPOWHEG+PYTHIA-output_had-max_Wpm.top      $dir
-
+    \cp pwgPOWHEG+PYTHIA-output_had-min_Wpm.top      $dir
+    \cp pwgPOWHEG+PYTHIA-output_had-max_Wpm.top      $dir
+    \cp pwgPOWHEG+PYTHIA-output_had-W1_Wpm.top       $dir/pwgPOWHEG+PYTHIA-output_had_Wpm.top
+fi
